# docs/swagger.yaml
basePath: /
definitions:
...
  swagger.Http400:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: Bad Request
        type: string
    type: object
  swagger.Http404:
    properties:
      code:
        example: 404
        type: integer
      msg:
        example: Page Not Found
        type: string
    type: object
  swagger.Http500:
    properties:
      code:
        example: 500
        type: integer
      msg:
        example: 服务器内部错误
        type: string
    type: object
  swagger.HttpOk:
    properties:
      code:
        example: 200
        type: integer
      msg:
        example: 接口调用成功
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: 这里写联系人信息
    url: https://www.swagger.io/support
  description: 这里写描述信息
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: 这里写标题
  version: "1.0"
paths:
...
  /user/UserLogin:
    post:
      consumes:
      - multipart/form-data
      description: Post 请求, 发送 Json 数据, 参数在消息体中 用户登陆
      parameters:
      - description: 登陆的用户名
        in: formData
        maxLength: 32
        minLength: 1
        name: username
        type: string
      - description: 用户密码
        in: formData
        maxLength: 32
        minLength: 1
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.HttpOk'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.Http400'
        "404":
          description: Page Not Found
          schema:
            $ref: '#/definitions/swagger.Http404'
        "500":
          description: InternalError
          schema:
            $ref: '#/definitions/swagger.Http500'
      summary: Post 请求, 发送 Json 数据, 参数在消息体中 用户登陆
      tags:
      - userApi
  /user/UserRegister:
    post:
      consumes:
      - multipart/form-data
      description: Post 请求, 发送 Json 数据, 参数在消息体中 用户注册
      parameters:
      - description: 要注册的用户名
        in: formData
        maxLength: 32
        minLength: 1
        name: username
        type: string
      - description: 用户密码
        in: formData
        maxLength: 32
        minLength: 1
        name: password
        type: string
      - description: 电话号码
        in: formData
        maxLength: 11
        minLength: 11
        name: tel
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.HttpOk'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.Http400'
        "404":
          description: Page Not Found
          schema:
            $ref: '#/definitions/swagger.Http404'
        "500":
          description: InternalError
          schema:
            $ref: '#/definitions/swagger.Http500'
      summary: Post 请求, 发送 Json 数据, 参数在消息体中 用户注册
      tags:
      - userApi
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"